Add navigation-bar.js module
