+++ modules/index.js
@@ -1,20 +1,15 @@
 import Book from './Book.js';
 import getTime from './date.js';
+import { listNavFun, addBookNavFun, contactSectionFun } from './navigation-bar.js';
 
 const bookListUi = document.querySelector('.book-list-ui');
 const addBtn = document.getElementById('add-btn');
 const bookTitle = document.querySelector('#title');
 const bookAuthor = document.querySelector('#author');
 
-const listNav = document.querySelector('.list');
-const addBookNav = document.querySelector('.add-list');
-const contactNav = document.querySelector('.contact');
-const contactSection = document.querySelector('.info-contact');
-
 const book = new Book();
 
 addBtn.addEventListener('click', () => {
-  const contactSection = document.querySelector('.info-contact');
   if (bookTitle.value.trim() === '' || bookAuthor.value.trim() === '') {
     return;
   }
@@ -23,27 +18,11 @@ addBtn.addEventListener('click', () => {
   bookAuthor.value = '';
   addBtn.parentElement.classList.add('hidden');
   bookListUi.parentElement.classList.remove('hidden');
-  contactSection.classList.add('hidden');
   book.renderBook();
 });
 
-listNav.addEventListener('click', () => {
-  bookListUi.parentElement.classList.remove('hidden');
-  addBtn.parentElement.classList.add('hidden');
-  contactSection.classList.add('hidden');
-});
-
-addBookNav.addEventListener('click', () => {
-  addBtn.parentElement.classList.remove('hidden');
-  bookListUi.parentElement.classList.add('hidden');
-  contactSection.classList.add('hidden');
-});
-
-contactNav.addEventListener('click', () => {
-  contactSection.classList.remove('hidden');
-  addBtn.parentElement.classList.add('hidden');
-  bookListUi.parentElement.classList.add('hidden');
-});
-
+listNavFun(addBtn, bookListUi);
+addBookNavFun(addBtn, bookListUi);
+contactSectionFun(addBtn, bookListUi);
 getTime();
 setInterval(getTime, 1000);
